+++ init/dx.h
@@ -68,6 +68,9 @@ typedef struct {
 	char *json;
 	int lines;
 	int json_parse_errors, dx_format_errors;
+	#define N_DX_FILE_HASH_BYTES 4      // 4 bytes = 8 chars
+    char file_hash[N_DX_FILE_HASH_BYTES*2 + SPACE_FOR_NULL];
+    int file_size;
 } dxlist_t;
 
 extern dxlist_t dx;
