+++ kiwi.config
@@ -96,27 +96,23 @@
 				// rx
 				// 12k/20.25k sample rates chosen to be integer multiples of WSPR 375 Hz sample rate
 			#if USE_RX_CICF
-				DEFp	RX1_WIDE_DECIM  823
-				DEFp	RX2_WIDE_DECIM  2			// 823*2 = 1646, 66.6666M/1646 = 40502.187 Hz (+2.187)
-				// DEFp	RX1_WIDE_DECIM  1543			// to get near integer multiple for WSPR sample rate
-				// DEFp	RX2_WIDE_DECIM  4			// 1543*4 = 6172, 125.0000M/6172 = 20252.754 Hz
+				DEFp	RX1_WIDE_DECIM  1543
+				DEFp	RX2_WIDE_DECIM  2			// 1543*2 = 3086, 125.0000M/3086 = 40502.187 Hz (+2.187)
 
-				DEFp	RX1_STD_DECIM   505			// to get near integer multiple for WSPR sample rate
-				DEFp	RX2_STD_DECIM   11			// 505*11 = 5555, 66.6666M/5555 = 12001.188 Hz
+				DEFp	RX1_STD_DECIM   1736
+				DEFp	RX2_STD_DECIM   3			// 1736*3 = 5209, 125.0000M/5209 = 23998.056 Hz (-1.944)
 				
 				DEFp    CICF_DECIM_BY_2 2
 			#else
-				DEFp	RX1_WIDE_DECIM  823
-				DEFp	RX2_WIDE_DECIM  4			// 823*4 = 3292, 66.6666M/3292 = 20251.094 Hz (+1.094)
-				// DEFp	RX1_STD_DECIM   947			// to get near integer multiple for WSPR sample rate
-				// DEFp	RX2_STD_DECIM   11			// 947*11 = 10417, 125.0000M/10417 = 11999.616 Hz
+				DEFp	RX1_WIDE_DECIM  1543
+				DEFp	RX2_WIDE_DECIM  4			// 1543*4 = 6173, 125.0000M/6173 = 20251.094 Hz (+1.094)
 				
 				// half as much error as below, but don't want to change status quo
 				//DEFp	RX1_STD_DECIM   926
 				//DEFp	RX2_STD_DECIM   6			// 505*11 = 5556, 66.6666M/5556 = 11999.028 Hz (-0.972)
 
-				DEFp	RX1_STD_DECIM   505
-				DEFp	RX2_STD_DECIM   11			// 505*11 = 5555, 66.6666M/5555 = 12001.188 Hz (+1.188)
+				DEFp	RX1_STD_DECIM   947
+				DEFp	RX2_STD_DECIM   11			// 505*11 = 10416, 125.0000M/10416 = 12001.188 Hz (+1.188)
 				
 				DEFp    CICF_DECIM_BY_2 1
             #endif
