+++ net/net.cpp
@@ -729,153 +729,6 @@ bool check_if_forwarded(const char *id, struct mg_connection *mc, char *remote_i
 }
 
 
-// updates net.ip_blacklist
-static int ip_blacklist_add(char *ips, bool *whitelist)
-{
-    char ip_str[NET_ADDRSTRLEN];
-    u4_t cidr;
-    *whitelist = false;
-    
-    if (ips[0] == '+') {
-        ips++;
-        *whitelist = true;
-    }
-    int n = sscanf(ips, "%[^/]/%d", ip_str, &cidr);
-    if (n == 0 || n > 2) return -1;
-    if (n == 1) cidr = 32;
-    bool error;
-    u1_t a,b,c,d;
-    u4_t ip = inet4_d2h(ip_str, &error, &a, &b, &c, &d);
-    if (error || cidr < 1 || cidr > 32) return -1;
-    u4_t nm = ~( (1 << (32-cidr)) -1 );
-    ip &= nm;       // make consistent with netmask
-    
-    if (ip == (net.ips_kiwisdr_com.ip[0] & nm)) {
-        lprintf("DANGER: blacklist entry %s would contain kiwisdr.com ip of %s, IGNORED!!!\n", ips, net.ips_kiwisdr_com.ip_list[0]);
-        return -1;
-    }
-    
-    int i = net.ip_blacklist_len;
-    if (i >= N_IP_BLACKLIST) {
-        lprintf("ip_blacklist_add: >= N_IP_BLACKLIST(%d)\n", N_IP_BLACKLIST);
-        return -1;
-    }
-    
-    // always add to beginning of list to match iptables insert behavior
-    if (net.ip_blacklist_len != 0)
-        memmove(&net.ip_blacklist[1], &net.ip_blacklist[0], net.ip_blacklist_len * sizeof(ip_blacklist_t));
-    ip_blacklist_t *bl = &net.ip_blacklist[0];
-    bl->ip = ip;
-    bl->a = a; bl->b = b; bl->c = c; bl->d = d;
-    bl->cidr = cidr;
-    bl->nm = nm;
-    bl->whitelist = *whitelist;
-    bl->dropped = bl->last_dropped = 0;
-    //printf("ip_blacklist_add[%d] %s %d.%d.%d.%d 0x%08x\n", net.ip_blacklist_len, ips, bl->a, bl->b, bl->c, bl->d, bl->nm);
-    net.ip_blacklist_len++;
-    
-    return 0;
-}
-
-// updates net.ip_blacklist (proxied Kiwis) and Linux iptables (non-proxied Kiwis)
-// called here and from admin "SET network_ip_blacklist="
-int ip_blacklist_add_iptables(char *ip_s)
-{
-    int rv;
-    //real_printf("    \"%s\",\n", ip_s);
-
-    #ifdef TEST_IP_BLACKLIST_USING_LOCAL_IPs
-    #else
-        bool is_loopback;
-        if (isLocal_ip(ip_s, &is_loopback) || is_loopback) {
-            lprintf("ip_blacklist_add_iptables: DANGER! local ip ignored: %s\n", ip_s);
-            return -1;
-        }
-    #endif
-    
-    bool whitelist;
-    if ((rv = ip_blacklist_add(ip_s, &whitelist)) != 0) return rv;
-
-    char *cmd_p;
-    // NB: insert (-I) NOT append (-A) because we may be incrementally adding after RETURN rule (last) exists
-    asprintf(&cmd_p, "iptables -I KIWI -s %s -j %s", ip_s, whitelist? "RETURN" : "DROP");
-    rv = non_blocking_cmd_system_child("kiwi.iptables", cmd_p, POLL_MSEC(200));
-    rv = WEXITSTATUS(rv);
-    lprintf("ip_blacklist_add_iptables: \"%s\" rv=%d\n", cmd_p, rv);
-    kiwi_ifree(cmd_p);
-    return rv;
-}
-
-void ip_blacklist_init_list(const char *list)
-{
-    const char *bl_s = admcfg_string(list, NULL, CFG_REQUIRED);
-    if (bl_s == NULL) return;
-
-    char *r_buf, *ips[N_IP_BLACKLIST+1];
-    int n = kiwi_split((char *) bl_s, &r_buf, " ", ips, N_IP_BLACKLIST);
-    //printf("ip_blacklist_init n=%d bl_s=\"%s\"\n", n, bl_s);
-    lprintf("ip_blacklist_init_list: %d entries: %s\n", n, list);
-    
-    for (int i=0; i < n; i++) {
-        ip_blacklist_add_iptables(ips[i]);
-    }
-
-    kiwi_ifree(r_buf);
-    admcfg_string_free(bl_s);
-}
-
-void ip_blacklist_init()
-{
-    net.ip_blacklist_len = 0;
-    // clean out old KIWI table first (if any)
-    system("iptables -D INPUT -j KIWI; iptables -F KIWI; iptables -X KIWI; iptables -N KIWI");
-    ip_blacklist_init_list("ip_blacklist");
-    ip_blacklist_init_list("ip_blacklist_local");
-    system("iptables -A KIWI -j RETURN; iptables -A INPUT -j KIWI");
-}
-
-// check internal blacklist for proxied Kiwis (iptables can't be used)
-bool check_ip_blacklist(char *remote_ip, bool log)
-{
-    bool error;
-    u4_t ip = inet4_d2h(remote_ip, &error);
-    if (error) return false;
-    for (int i=0; i < net.ip_blacklist_len; i++) {
-        ip_blacklist_t *bl = &net.ip_blacklist[i];
-        u4_t nm = bl->nm;
-        if ((ip & nm) == bl->ip) {      // netmask previously applied to bl->ip
-            net.ip_blacklist_inuse = true;
-            bl->dropped++;
-            if (bl->whitelist) return false;
-            bl->last_dropped = ip;
-            if (log) lprintf("IP BLACKLISTED: %s\n", remote_ip);
-            return true;
-        }
-    }
-    return false;
-}
-
-void ip_blacklist_dump()
-{
-    //if (!net.ip_blacklist_inuse) return;
-	lprintf("\n");
-	lprintf("PROXY IP BLACKLIST:\n");
-	lprintf("  dropped  ip\n");
-	
-	for (int i = 0; i < net.ip_blacklist_len; i++) {
-	    ip_blacklist_t *bl = &net.ip_blacklist[i];
-	    //if (bl->dropped == 0) continue;
-	    u1_t a, b, c, d;
-	    inet4_h2d(bl->last_dropped, &a, &b, &c, &d);
-	    lprintf("%9d  %18s %08x|%08x last=%d.%d.%d.%d|%08x %s\n",
-	        bl->dropped, stprintf("%d.%d.%d.%d/%d", bl->a, bl->b, bl->c, bl->d, bl->cidr),
-	        bl->ip, bl->nm, a, b, c, d, bl->last_dropped,
-	        bl->whitelist? " WHITELIST" : "");
-	}
-	lprintf("\n");
-}
-
-
 // Emulates the client-side (js) of a Kiwi sound/waterfall/extension channel API connection.
 // For use by server-side internal connections, e.g. WSPR autorun, FT8 autorun, SNR measurement.
 
@@ -955,8 +808,8 @@ bool internal_conn_setup(u4_t ws, internal_conn_t *iconn, int instance, int port
     return true;
 
 error:
-    printf("internal_conn_setup: %s couldn't get websocket instance=%d uri=%s port=%d\n",
-        ident_user, instance, mc_fail->uri, mc_fail->remote_port);
+    //printf("internal_conn_setup: %s couldn't get websocket instance=%d uri=%s port=%d\n",
+    //    ident_user, instance, mc_fail->uri, mc_fail->remote_port);
     internal_conn_shutdown(iconn);
     return false;
