+++ rx/rx_server_ajax.cpp
@@ -443,12 +443,15 @@ fail:
         u4_t level = 0;
         // "%i" so decimal or hex beginning with 0x can be specified
         if (mc->query_string && sscanf(mc->query_string, "level=%i", &level) == 1) {
-            adc_level = level & ((1 << ADC_BITS) - 1);
+            adc_level = level & ((1 << (ADC_BITS-1)) - 1);
             //printf("/adc SET level=%d(0x%x)\n", adc_level, adc_level);
+            #define COUNT_ADC_OVFL 0x2000
+            if (adc_level == 0) adc_level = COUNT_ADC_OVFL;     // count ADC overflow instead of a level value
             spi_set(CmdSetADCLvl, adc_level);
         }
-		asprintf(&sb, "{ \"adc_level_dec\":%u, \"adc_level_hex\":\"0x%x\", \"adc_count\":%u }\n",
-		    adc_level, adc_level, adc_count);
+        u4_t _adc_level = (adc_level == COUNT_ADC_OVFL)? 0 : adc_level;
+		asprintf(&sb, "{ \"adc_level_dec\":%u, \"adc_level_hex\":\"0x%x\", \"adc_count\":%u, \"ver_maj\":%d, \"ver_min\":%d }\n",
+		    _adc_level, _adc_level, adc_count, version_maj, version_min);
 		break;
 	}
 
@@ -541,6 +544,11 @@ fail:
 			status = "offline";
 		else
 			status = "active";
+		
+		// number of channels preemptible, but only if external preemption enabled
+		int preempt = 0;
+		if (any_preempt_autorun)
+		    rx_chan_free_count(RX_COUNT_ALL, NULL, NULL, &preempt);
 
 		// the avatar file is in the in-memory store, so it's not going to be changing after server start
 		u4_t avatar_ctime = timer_server_build_unix_time();
@@ -561,7 +569,8 @@ fail:
 			"name=%s\nsdr_hw=KiwiSDR v%d.%d%s%s%s%s%s%s%s%s ⁣\n"
 			"op_email=%s\n"
 			"bands=%.0f-%.0f\nfreq_offset=%.3f\n"
-			"users=%d\nusers_max=%d\navatar_ctime=%u\n"
+			"users=%d\nusers_max=%d\npreempt=%d\n"
+			"avatar_ctime=%u\n"
 			"gps=%s\ngps_good=%d\nfixes=%d\nfixes_min=%d\nfixes_hour=%d\n"
 			"tdoa_id=%s\ntdoa_ch=%d\n"
 			"asl=%d\n"
@@ -575,7 +584,8 @@ fail:
 			"gps_date=%d,%d\n"
 			"date=%s\n"
 			//"test=7\n"
-			"ip_blacklist=%s\n",
+			"ip_blacklist=%s\n"
+			"dx_file=%d,%s,%d\n",
 			
 			status, no_open_access? "auth=password\n" : "", offline? "yes":"no",
 			name, version_maj, version_min,
@@ -603,7 +613,8 @@ fail:
 
 			(s3 = cfg_string("admin_email", NULL, CFG_OPTIONAL)),
 			(float) kiwi_reg_lo_kHz * kHz, (float) kiwi_reg_hi_kHz * kHz, freq_offset_kHz,
-			users, users_max, avatar_ctime,
+			users, users_max, preempt,
+			avatar_ctime,
 			gps_loc,
 			#ifdef USE_GPS
 			    gps.good, gps.fixes, gps.fixes_min, gps.fixes_hour,
@@ -625,7 +636,8 @@ fail:
 			    0, 0,
 			#endif
 			utc_ctime_static(),
-			net.ip_blacklist_hash
+			net.ip_blacklist_hash,
+			dx.stored_len, dx.file_hash, dx.file_size
 			);
 
 		kiwi_ifree(name);
