+++ verilog.Vivado.2022.2.ip/README
@@ -575,7 +575,7 @@ used by rx/{iq_mixer.v, gen.v}:
             > phase generator and sin cos lut
         System Requirements
             System Clock
-                > 66.6666
+                > 125.0000
             Number of Channels
                 > 1
             Mode of Operation
@@ -631,8 +631,8 @@ used by rx/{iq_mixer.v, gen.v}:
 	Summary (if anything here is different you configured something incorrectly)
         Output Width = 15 Bits
         Channels = 1
-        System Clock = 66.6666 MHz
-        Frequency per Channel (Fs) = 66.666600000000003 MHz
+        System Clock = 125.0000 MHz
+        Frequency per Channel (Fs) = 125.0 MHz
         Noise Shaping = Phase Dithering
         Memory Type = Block ROM
         Optimization Goal = Area
@@ -652,7 +652,7 @@ used by rx/iq_mixer_48b.v:
             > phase generator and sin cos lut
         System Requirements
             System Clock
-                > 66.6666
+                > 125.0000
             Number of Channels
                 > 1
             Mode of Operation
@@ -708,8 +708,8 @@ used by rx/iq_mixer_48b.v:
 	Summary (if anything here is different you configured something incorrectly)
         Output Width = 15 Bits
         Channels = 1
-        System Clock = 66.6666 MHz
-        Frequency per Channel (Fs) = 66.666600000000003 MHz
+        System Clock = 125.0000 MHz
+        Frequency per Channel (Fs) = 125.0 MHz
         Noise Shaping = Phase Dithering
         Memory Type = Block ROM
         Optimization Goal = Area
