+++ web/extensions/FT8/FT8.js
@@ -298,7 +298,10 @@ function FT8_config_html()
                   '</h6></header>'
                ),
                
-               w3_button('id-ft8-restart w3-margin-T-16 w3-aqua w3-hide', 'autorun restart', 'ft8_autorun_restart_cb'),
+               w3_inline('w3-margin-top/',
+                  w3_button('w3-blue', 'set all to regular use', 'ft8_autorun_all_regular_cb'),
+                  w3_button('id-ft8-restart w3-margin-left w3-aqua w3-hide', 'autorun restart', 'ft8_autorun_restart_cb')
+               ),
                
                w3_div('id-ft8-admin-autorun w3-margin-T-16')
             )
@@ -347,8 +350,22 @@ function ft8_autorun_restart_cb()
 
 function ft8_autorun_select_cb(path, idx, first)
 {
+   //console.log('ft8_autorun_select_cb: path='+ path +' idx='+ idx +' first='+ first);
    admin_select_cb(path, idx, first);
    if (first) return;
+   w3_show('id-ft8-restart');
+	var el = w3_el('id-kiwi-container');
+	w3_scrollDown(el);   // keep menus visible
+}
+
+function ft8_autorun_all_regular_cb(path, idx, first)
+{
+   if (first) return;
+   for (var i = 0; i < rx_chans; i++) {
+      var path = 'ft8.autorun'+ i;
+      w3_select_value(path, 0);
+      admin_select_cb(path, 0);
+   }
    w3_show('id-ft8-restart');
 	var el = w3_el('id-kiwi-container');
 	w3_scrollDown(el);   // keep menus visible
